Created new file "info_search" to handle different types of html formats.
Created list of possible classes that are found as well as a fallback block to be able to search for key words in HTML.

Brought in LSU again for testing of table data found in its HTML.

Tested with UVA space's HTML to find commonalities and bugs.

Created multiple chances in loop in test.py to disqualify certain blocks if they did not contain desired content in each. (Line 33 / 40-41 / 49-50 / 53 / 64-65.